text2image-stable-diffusion/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
diffusers==0.33.1
fastapi==0.115.12
filelock==3.13.1
fsspec==2024.6.1
h11==0.16.0
huggingface-hub==0.31.4
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.2.6
packaging==25.0
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
setuptools==70.2.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.3
tokenizers==0.21.1
torch==2.7.0+cpu
tqdm==4.67.1
transformers==4.52.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
zipp==3.21.0
